move `TodoList.tsx` of `react-drizzle` to `components` dir
